+++ src/components/Pagination.tsx
@@ -1,20 +1,21 @@
+import { useParams, useNavigate, useSearchParams } from "react-router-dom";
 import { HandlerButtonType } from "../types/types";
 
 interface IPagination {
   pageNumber: number;
-  handlePrev: () => void;
-  handleNext: () => void;
-  paginate: HandlerButtonType;
+
   pageIndex: number;
 }
 
 const Pagination = ({
   pageNumber,
-  paginate,
-  handlePrev,
-  handleNext,
+
   pageIndex,
 }: IPagination) => {
+  const [searchParams, setSearchParams] = useSearchParams();
+  const { page } = useParams();
+  console.log(page);
+  const navigate = useNavigate();
   let pages = [];
   if (pageNumber) {
     pages = new Array(pageNumber).fill(0).map((item, index) => {
@@ -22,6 +23,34 @@ const Pagination = ({
       return item;
     });
   }
+
+  const handlePagination: HandlerButtonType = (event) => {
+    const target = event.target as HTMLElement;
+    if (target.textContent) {
+      const lastQuery: string | null = localStorage.getItem("lastQuery");
+      searchParams.set("name", lastQuery ? lastQuery : "");
+      setSearchParams(searchParams);
+      navigate(`/${+target.textContent}` + "?" + searchParams);
+    }
+  };
+  const handlePrev = () => {
+    if (page) {
+      if (+page > 1) {
+        navigate(`/${+page - 1}` + "?" + searchParams);
+      }
+    } else {
+      navigate(`/`);
+    }
+  };
+  const handleNext = () => {
+    if (page) {
+      if (+page < pageNumber) {
+        navigate(`/${+page + 1}` + "?" + searchParams);
+      }
+    } else {
+      navigate(`/`);
+    }
+  };
   return (
     <ul className="container mx-auto flex justify-center flex-wrap gap-3 p-[20px]">
       <button onClick={handlePrev} className="cursor-pointer hover:opacity-60">
@@ -47,7 +76,7 @@ const Pagination = ({
             ` cursor-pointer p-1.5 rounded`
           }
           key={(Date.now() * Math.random()).toString()}
-          onClick={paginate}
+          onClick={handlePagination}
         >
           {item}
         </li>
